requirements.txt
fastapi
uvicorn
argon2-cffi
pydantic
aiofiles